Complete "Creating an invoice" part of chapter 12
There is a bug when adding an invoice with an amount in $ and cents: the conversion in cents does not seem to work and the number is not coerced to an integer as it should. The tutorial does not mention anything about it. It might be a bug with the "zod" library, that is supposed to do the job. It is not crucial to continue with the tutorial since adding invoices with integer amounts ($ with no cents) works fine.